refactor(tests): migrate all visual tests to new fixture system
- Updated all test files to import from new centralized fixtures:
  - character-creation.spec.ts
  - ending-screen.spec.ts
  - game-session.spec.ts
  - journal-modal.spec.ts
  - main-pages.spec.ts
  - world-creation-ai-guidance.spec.ts
  - world-creation.spec.ts
  - global.setup.ts

- Deleted backward compatibility layer (data-seeder.ts)
- All tests now import directly from:
  - seedTestData utility (./utils/seedTestData)
  - mockApi utility (./utils/mockApi)
  - Fixture data (../fixtures/*.fixture)

This completes the migration from the old inline test data approach
to the modern centralized fixture system with factory patterns.
