+++ tests/visual/ending-screen.spec.ts
@@ -1,6 +1,7 @@
 import { test, expect } from '@playwright/test';
 import { waitForContentStable, hideDynamicContent } from './utils/wait-helpers';
-import { seedTestData, mockApiEndpoints } from './utils/data-seeder';
+import { seedTestData } from './utils/seedTestData';
+import { mockApiEndpoints } from './utils/mockApi';
 
 /**
  * EndingScreen Visual Regression Tests
