+++ tests/visual/world-creation-ai-guidance.spec.ts
@@ -1,6 +1,7 @@
 import { test, expect } from '@playwright/test';
 import { waitForContentStable } from './utils/wait-helpers';
-import { seedTestData, mockApiEndpoints } from './utils/data-seeder';
+import { seedTestData } from './utils/seedTestData';
+import { mockApiEndpoints } from './utils/mockApi';
 
 test.describe('World Creation Wizard AI Guidance', () => {
   test.setTimeout(45000);
